Consolidate ParsedInputAddress into model project

+++ geocoder/src/main/scala/grasshopper/geocoder/api/GeocodeFlow.scala
@@ -3,12 +3,11 @@ package grasshopper.geocoder.api
 import akka.stream.scaladsl._
 import grasshopper.client.addresspoints.AddressPointsClient
 import grasshopper.client.census.CensusClient
-import grasshopper.client.census.model.ParsedInputAddress
 import grasshopper.client.parser.AddressParserClient
 import grasshopper.client.parser.model.ParsedAddress
 import grasshopper.geocoder.model.{ BatchGeocodeResult, AddressPointsGeocodeBatchResult, CensusGeocodeBatchResult, ParsedOutputBatchAddress }
 import grasshopper.model.addresspoints.AddressPointsResult
-import grasshopper.model.census.CensusResult
+import grasshopper.model.census.{ ParsedInputAddress, CensusResult }
 
 import scala.concurrent.ExecutionContext
 
@@ -32,7 +31,7 @@ object GeocodeFlows {
         ParsedOutputBatchAddress(
           a.input,
           ParsedInputAddress(
-            a.parts.addressNumber.toInt,
+            a.parts.addressNumber,
             a.parts.streetName,
             a.parts.zip,
             a.parts.state
