+++ geocoder/src/main/scala/grasshopper/geocoder/api/Service.scala
@@ -17,14 +17,13 @@ import com.typesafe.scalalogging.Logger
 import grasshopper.model.Status
 import grasshopper.client.addresspoints.AddressPointsClient
 import grasshopper.client.census.CensusClient
-import grasshopper.client.census.model.ParsedInputAddress
 import grasshopper.client.parser.AddressParserClient
 import grasshopper.client.parser.model.{ ParsedAddress, ParserStatus }
 import grasshopper.client.protocol.ClientJsonProtocol
 import grasshopper.geocoder.model.{ AddressPointsGeocodeBatchResult, CensusGeocodeBatchResult, GeocodeResult, GeocodeStatus }
 import grasshopper.geocoder.protocol.GrasshopperJsonProtocol
 import grasshopper.model.addresspoints.AddressPointsResult
-import grasshopper.model.census.CensusResult
+import grasshopper.model.census.{ ParsedInputAddress, CensusResult }
 import org.slf4j.LoggerFactory
 
 import scala.async.Async.{ async, await }
@@ -105,7 +104,7 @@ trait Service extends GrasshopperJsonProtocol with ClientJsonProtocol {
           } else {
             val parsedAddress = addr.right.getOrElse(ParsedAddress.empty)
             val parsedInputAddress = ParsedInputAddress(
-              parsedAddress.parts.addressNumber.toInt,
+              parsedAddress.parts.addressNumber,
               parsedAddress.parts.streetName,
               parsedAddress.parts.zip,
               parsedAddress.parts.state
